+++ GoedeDoelenHelpen/Pages/Charities/Delete.cshtml
@@ -1,12 +1,13 @@
 ﻿@page
-@model GoedeDoelenHelpen.Pages.Charity.DetailsModel
+@model GoedeDoelenHelpen.Pages.Charities.DeleteModel
 
 @{
-    ViewData["Title"] = "Details";
+    ViewData["Title"] = "Delete";
 }
 
-<h2>Details</h2>
+<h2>Delete</h2>
 
+<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Charity</h4>
     <hr />
@@ -66,8 +67,10 @@
             @Html.DisplayFor(model => model.Charity.IBAN)
         </dd>
     </dl>
-</div>
-<div>
-    <a asp-page="./Edit" asp-route-id="@Model.Charity.Id">Edit</a> |
-    <a asp-page="./Index">Back to List</a>
+    
+    <form method="post">
+        <input type="hidden" asp-for="Charity.Id" />
+        <input type="submit" value="Delete" class="btn btn-default" /> |
+        <a asp-page="./Index">Back to List</a>
+    </form>
 </div>
