+++ GoedeDoelenHelpen/Pages/Charities/Delete.cshtml.cs
@@ -7,18 +7,19 @@ using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using GoedeDoelenHelpen.Data;
 
-namespace GoedeDoelenHelpen.Pages.Charity
+namespace GoedeDoelenHelpen.Pages.Charities
 {
-    public class DetailsModel : PageModel
+    public class DeleteModel : PageModel
     {
         private readonly GoedeDoelenHelpen.Data.ApplicationDbContext _context;
 
-        public DetailsModel(GoedeDoelenHelpen.Data.ApplicationDbContext context)
+        public DeleteModel(GoedeDoelenHelpen.Data.ApplicationDbContext context)
         {
             _context = context;
         }
 
-        public Data.Charity Charity { get; set; }
+        [BindProperty]
+        public Charity Charity { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
@@ -34,6 +35,24 @@ namespace GoedeDoelenHelpen.Pages.Charity
                 return NotFound();
             }
             return Page();
+        }
+
+        public async Task<IActionResult> OnPostAsync(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            Charity = await _context.Charities.FindAsync(id);
+
+            if (Charity != null)
+            {
+                _context.Charities.Remove(Charity);
+                await _context.SaveChangesAsync();
+            }
+
+            return RedirectToPage("./Index");
         }
     }
 }
