+++ GoedeDoelenHelpen/Pages/Charities/Details.cshtml
@@ -1,13 +1,12 @@
 ﻿@page
-@model GoedeDoelenHelpen.Pages.Charity.DeleteModel
+@model GoedeDoelenHelpen.Pages.Charities.DetailsModel
 
 @{
-    ViewData["Title"] = "Delete";
+    ViewData["Title"] = "Details";
 }
 
-<h2>Delete</h2>
+<h2>Details</h2>
 
-<h3>Are you sure you want to delete this?</h3>
 <div>
     <h4>Charity</h4>
     <hr />
@@ -67,10 +66,8 @@
             @Html.DisplayFor(model => model.Charity.IBAN)
         </dd>
     </dl>
-    
-    <form method="post">
-        <input type="hidden" asp-for="Charity.Id" />
-        <input type="submit" value="Delete" class="btn btn-default" /> |
-        <a asp-page="./Index">Back to List</a>
-    </form>
+</div>
+<div>
+    <a asp-page="./Edit" asp-route-id="@Model.Charity.Id">Edit</a> |
+    <a asp-page="./Index">Back to List</a>
 </div>
