+++ GoedeDoelenHelpen/Pages/Charities/Edit.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model GoedeDoelenHelpen.Pages.Charity.EditModel
+@model GoedeDoelenHelpen.Pages.Charities.EditModel
 
 @{
     ViewData["Title"] = "Edit";
@@ -15,6 +15,11 @@
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="Charity.Id" />
             <div class="form-group">
+                <label asp-for="Charity.Name" class="control-label"></label>
+                <input asp-for="Charity.Name" class="form-control" />
+                <span asp-validation-for="Charity.Name" class="text-danger"></span>
+            </div>
+            <div class="form-group">
                 <label asp-for="Charity.Adress" class="control-label"></label>
                 <input asp-for="Charity.Adress" class="form-control" />
                 <span asp-validation-for="Charity.Adress" class="text-danger"></span>
@@ -29,11 +34,7 @@
                 <input asp-for="Charity.City" class="form-control" />
                 <span asp-validation-for="Charity.City" class="text-danger"></span>
             </div>
-            <div class="form-group">
-                <label asp-for="Charity.Name" class="control-label"></label>
-                <input asp-for="Charity.Name" class="form-control" />
-                <span asp-validation-for="Charity.Name" class="text-danger"></span>
-            </div>
+            <hr />
             <div class="form-group">
                 <label asp-for="Charity.KVKNumber" class="control-label"></label>
                 <input asp-for="Charity.KVKNumber" class="form-control" />
@@ -47,7 +48,7 @@
             <div class="form-group">
                 <div class="checkbox">
                     <label>
-                        <input asp-for="Charity.Activated" /> @Html.DisplayNameFor(model => model.Activated)
+                        <input asp-for="Charity.Activated" /> @Html.DisplayNameFor(model => model.Charity.Activated)
                     </label>
                 </div>
             </div>
