+++ GoedeDoelenHelpen/Pages/Charities/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model GoedeDoelenHelpen.Pages.Charity.IndexModel
+@model GoedeDoelenHelpen.Pages.Charities.IndexModel
 
 @{
     ViewData["Title"] = "Index";
